microservices/permission/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
PyYAML
pydantic-settings 
httpx
python-consul